perf(startup): slim boot critical path and vendor chunking
- viteManualChunks: isolate react/react-dom/scheduler into framework
  chunk so feature vendor chunks stop becoming static entry deps
  (root cause of tiptap 533KB being force-preloaded at boot)
- main.tsx: drop eager katex css, DEV-only scrollPerfDebug import
- markdownMath: load katex css on first markdown processor attach
- ProjectRulesModal / SideTabBody: lazy MarkdownTiptapEditor / TerminalTab
- OfficeDocumentPreview: split into pdf/docx/xlsx lazy sub-previews
- CodePreview: trim highlight.js registrations 68 -> 22 (preview falls
  back to plain text; codeLang map kept full for CodeMirror editor)
- delete dead ResourceViewer shell + 5 orphan modules (-4717 lines)
  and unused dompurify dependency

Fixes #1055
