+++ package.json
@@ -73,7 +73,6 @@
     "@xterm/xterm": "^6.0.0",
     "clsx": "^2.1.1",
     "docx-preview": "^0.4.0",
-    "dompurify": "^3.4.13",
     "highlight.js": "^11.11.1",
     "html-to-image": "^1.11.13",
     "katex": "^0.18.4",
@@ -101,7 +100,6 @@
     "@testing-library/jest-dom": "^7.0.1",
     "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.6.5",
-    "@types/dompurify": "^3.2.0",
     "@types/node": "^22.15.0",
     "@types/qrcode": "^1.5.6",
     "@types/react": "^19.1.8",
