+++ src/components/OfficeDocumentPreview.tsx
@@ -1,15 +1,15 @@
 /**
- * Rich local document preview:
- * - PDF  → react-pdf (pdf.js)
- * - DOCX → docx-preview (styled Word layout)
- * - XLSX → SheetJS (xlsx ≥0.20.3) multi-sheet tables
+ * Rich local document preview — dispatcher.
+ * - PDF  → OfficePdfPreview (react-pdf, lazy)
+ * - DOCX → OfficeDocxPreview (docx-preview, lazy)
+ * - XLSX → OfficeXlsxPreview (SheetJS, lazy)
  * - PPTX → limited text fallback + open externally
+ *
+ * Fetching / error / fallback chrome lives here so each format's renderer
+ * only ships (and parses) when that file type is actually opened.
  */
 
-import { useEffect, useMemo, useRef, useState } from "react";
-import { Document, Page, pdfjs } from "react-pdf";
-import { renderAsync } from "docx-preview";
-import * as XLSX from "xlsx";
+import { lazy, Suspense, useCallback, useEffect, useMemo, useState } from "react";
 import { fetchPreviewArrayBuffer } from "@/lib/filePreviewSrc";
 import {
   formatMediaLoadErrorMessage,
@@ -17,15 +17,19 @@ import {
 } from "@/lib/mediaLoadPro";
 import { createT, type Locale } from "@/i18n";
 import { openInEditor, pathOpen, pathReveal } from "@/lib/api";
-import { sanitizeOfficeSheetHtml } from "@/lib/sanitizeOfficeHtml";
-import { Tip } from "@/components/ui/tooltip";
-import { runAfterPaneSplitMotion } from "@/lib/paneSplitMotion";
-import pdfWorkerUrl from "pdfjs-dist/build/pdf.worker.min.mjs?url";
 
-import "react-pdf/dist/Page/AnnotationLayer.css";
-import "react-pdf/dist/Page/TextLayer.css";
-
-pdfjs.GlobalWorkerOptions.workerSrc = pdfWorkerUrl;
+const OfficePdfPreview = lazy(async () => {
+  const m = await import("@/components/OfficePdfPreview");
+  return { default: m.OfficePdfPreview };
+});
+const OfficeDocxPreview = lazy(async () => {
+  const m = await import("@/components/OfficeDocxPreview");
+  return { default: m.OfficeDocxPreview };
+});
+const OfficeXlsxPreview = lazy(async () => {
+  const m = await import("@/components/OfficeXlsxPreview");
+  return { default: m.OfficeXlsxPreview };
+});
 
 export interface OfficeDocumentPreviewProps {
   kind: string;
@@ -59,44 +63,10 @@ export function OfficeDocumentPreview({
 }: OfficeDocumentPreviewProps) {
   const tr = useMemo(() => createT(locale), [locale]);
   const [load, setLoad] = useState<LoadState>({ status: "loading" });
-  const [pdfPages, setPdfPages] = useState(0);
-  const [pdfPage, setPdfPage] = useState(1);
-  const [pdfScale, setPdfScale] = useState(1.05);
-  const [sheetNames, setSheetNames] = useState<string[]>([]);
-  const [activeSheet, setActiveSheet] = useState(0);
-  const [sheetHtml, setSheetHtml] = useState("");
-  const docxRef = useRef<HTMLDivElement>(null);
-  const docxScrollRef = useRef<HTMLDivElement>(null);
-
-  /**
-   * Force pages to use the pane width so text/tables reflow instead of
-   * clipping (docx-preview writes fixed page widths as inline styles).
-   */
-  const relaxDocxPageWidths = () => {
-    const host = docxRef.current;
-    if (!host) return;
-    host.querySelectorAll<HTMLElement>("section.docx").forEach((sec) => {
-      sec.style.setProperty("width", "100%", "important");
-      sec.style.setProperty("max-width", "100%", "important");
-      sec.style.setProperty("min-width", "0", "important");
-      sec.style.setProperty("box-sizing", "border-box", "important");
-    });
-    const wrap = host.querySelector<HTMLElement>(".docx-wrapper");
-    if (wrap) {
-      wrap.style.setProperty("width", "100%", "important");
-      wrap.style.setProperty("max-width", "100%", "important");
-      wrap.style.setProperty("padding", "0", "important");
-    }
-  };
 
   useEffect(() => {
     let cancelled = false;
     setLoad({ status: "loading" });
-    setPdfPages(0);
-    setPdfPage(1);
-    setSheetNames([]);
-    setActiveSheet(0);
-    setSheetHtml("");
 
     if (errorFromHost) {
       const resolved = resolveMediaLoadError(errorFromHost, "office");
@@ -143,122 +113,7 @@ export function OfficeDocumentPreview({
     };
   }, [absolutePath, kind, errorFromHost, tr]);
 
-  // DOCX render — reflow to pane width (full text, no side clip)
-  useEffect(() => {
-    if (load.status !== "ready") return;
-    if (kind !== "docx" && kind !== "office") return;
-    const el = docxRef.current;
-    if (!el) return;
-    el.innerHTML = "";
-    el.style.zoom = "";
-    el.style.transform = "";
-    let cancelled = false;
-    let ro: ResizeObserver | null = null;
-
-    void renderAsync(load.buffer, el, undefined, {
-      className: "office-docx-body",
-      inWrapper: true,
-      // Critical: ignore fixed page width so content uses the container
-      // (otherwise Chinese titles / tables overflow and get clipped).
-      ignoreWidth: true,
-      ignoreHeight: true,
-      breakPages: true,
-      renderHeaders: true,
-      renderFooters: true,
-      renderFootnotes: true,
-      useBase64URL: true,
-      experimental: true,
-    })
-      .then(() => {
-        if (cancelled) return;
-        relaxDocxPageWidths();
-        requestAnimationFrame(() => {
-          if (!cancelled) relaxDocxPageWidths();
-        });
-        // Images can change layout after load
-        el.querySelectorAll("img").forEach((img) => {
-          if (img.complete) return;
-          img.addEventListener(
-            "load",
-            () => {
-              if (!cancelled) relaxDocxPageWidths();
-            },
-            { once: true },
-          );
-        });
-        const scroll = docxScrollRef.current;
-        if (scroll && typeof ResizeObserver !== "undefined") {
-          ro = new ResizeObserver(() => {
-            if (cancelled) return;
-            if (runAfterPaneSplitMotion(relaxDocxPageWidths)) return;
-            relaxDocxPageWidths();
-          });
-          ro.observe(scroll);
-        }
-      })
-      .catch((e) => {
-        if (!cancelled) {
-          setLoad({
-            status: "error",
-            message: e instanceof Error ? e.message : String(e),
-          });
-        }
-      });
-    return () => {
-      cancelled = true;
-      ro?.disconnect();
-    };
-  }, [load, kind]);
-
-  // XLSX parse
-  useEffect(() => {
-    if (load.status !== "ready") return;
-    if (kind !== "xlsx") return;
-    try {
-      const wb = XLSX.read(load.buffer, { type: "array" });
-      const names = wb.SheetNames;
-      setSheetNames(names);
-      const idx = 0;
-      setActiveSheet(idx);
-      const ws = wb.Sheets[names[idx]];
-      setSheetHtml(
-        ws
-          ? sanitizeOfficeSheetHtml(
-              XLSX.utils.sheet_to_html(ws, { id: "office-sheet" }),
-            )
-          : "",
-      );
-    } catch (e) {
-      setLoad({
-        status: "error",
-        message: e instanceof Error ? e.message : String(e),
-      });
-    }
-  }, [load, kind]);
-
-  const switchSheet = (idx: number) => {
-    if (load.status !== "ready" || kind !== "xlsx") return;
-    try {
-      const wb = XLSX.read(load.buffer, { type: "array" });
-      const name = wb.SheetNames[idx];
-      const ws = wb.Sheets[name];
-      setActiveSheet(idx);
-      setSheetHtml(
-        ws
-          ? sanitizeOfficeSheetHtml(
-              XLSX.utils.sheet_to_html(ws, { id: "office-sheet" }),
-            )
-          : "",
-      );
-    } catch (e) {
-      setLoad({
-        status: "error",
-        message: e instanceof Error ? e.message : String(e),
-      });
-    }
-  };
-
-  const openExternal = async () => {
+  const openExternal = useCallback(async () => {
     try {
       await pathOpen(absolutePath);
     } catch {
@@ -268,18 +123,11 @@ export function OfficeDocumentPreview({
         await pathReveal(absolutePath);
       }
     }
-  };
+  }, [absolutePath]);
 
-  /**
-   * Stable react-pdf `file` prop. Inline `new Uint8Array(...)` every render
-   * remounts Document in a tight loop (GPU thrash → full-window black freeze
-   * when users open a generated PDF from chat).
-   */
-  const pdfBuffer = load.status === "ready" ? load.buffer : null;
-  const pdfFile = useMemo(() => {
-    if (!pdfBuffer || kind !== "pdf") return null;
-    return { data: new Uint8Array(pdfBuffer) };
-  }, [pdfBuffer, kind]);
+  const handleChildError = useCallback((message: string) => {
+    setLoad({ status: "error", message });
+  }, []);
 
   if (load.status === "loading") {
     return (
@@ -312,197 +160,64 @@ export function OfficeDocumentPreview({
     );
   }
 
-  // PDF — page/zoom chrome only; filename/open live in host when embedded
-  if (kind === "pdf") {
-    return (
-      <div className="office-preview office-preview--pdf">
-        <div className="office-preview__bar office-preview__bar--controls">
-          {!embedded ? (
-            <Tip label={name}>
-              <span className="office-preview__bar-title">
-                {name}
-              </span>
-            </Tip>
-          ) : (
-            <span className="office-preview__bar-spacer" />
-          )}
-          <div className="office-preview__bar-actions">
-            <button
-              type="button"
-              className="btn btn--ghost btn--sm"
-              disabled={pdfPage <= 1}
-              onClick={() => setPdfPage((p) => Math.max(1, p - 1))}
-            >
-              {tr("office.prevPage")}
-            </button>
-            <span className="office-preview__page">
-              {pdfPages
-                ? tr("office.pageOf", { page: pdfPage, total: pdfPages })
-                : "—"}
-            </span>
-            <button
-              type="button"
-              className="btn btn--ghost btn--sm"
-              disabled={!pdfPages || pdfPage >= pdfPages}
-              onClick={() =>
-                setPdfPage((p) => (pdfPages ? Math.min(pdfPages, p + 1) : p))
-              }
-            >
-              {tr("office.nextPage")}
-            </button>
-            <button
-              type="button"
-              className="btn btn--ghost btn--sm"
-              onClick={() => setPdfScale((s) => Math.max(0.6, s - 0.1))}
-            >
-              −
-            </button>
-            <button
-              type="button"
-              className="btn btn--ghost btn--sm"
-              onClick={() => setPdfScale((s) => Math.min(2.2, s + 0.1))}
-            >
-              +
-            </button>
-            {!embedded && (
-              <button
-                type="button"
-                className="btn btn--ghost btn--sm"
-                onClick={() => void openExternal()}
-              >
-                {tr("office.openExternal")}
-              </button>
-            )}
-          </div>
-        </div>
-        <div className="office-preview__pdf-scroll">
-          {pdfFile ? (
-            <Document
-              file={pdfFile}
-              onLoadSuccess={(d) => {
-                setPdfPages(d.numPages);
-                setPdfPage(1);
-              }}
-              loading={
-                <div className="office-preview__status">
-                  {tr("office.loading")}
-                </div>
-              }
-              error={
-                <div className="office-preview__status">
-                  {tr("office.renderFailed")}
-                </div>
-              }
-            >
-              <Page
-                pageNumber={pdfPage}
-                scale={pdfScale}
-                renderTextLayer
-                renderAnnotationLayer
-              />
-            </Document>
-          ) : null}
-        </div>
-      </div>
-    );
-  }
-
-  // DOCX — pure document body when embedded (no filename bar)
-  if (kind === "docx" || kind === "office") {
-    return (
-      <div
-        className={
-          "office-preview office-preview--docx" +
-          (embedded ? " office-preview--embedded" : "")
-        }
-      >
-        {!embedded && (
-          <div className="office-preview__bar">
-            <Tip label={name}>
-              <span className="office-preview__bar-title">
-                {name}
-              </span>
-            </Tip>
-            <div className="office-preview__bar-actions">
-              <button
-                type="button"
-                className="btn btn--ghost btn--sm"
-                onClick={() => void openExternal()}
-              >
-                {tr("office.openExternal")}
-              </button>
-            </div>
-          </div>
-        )}
-        <div
-          ref={docxScrollRef}
-          className="office-preview__docx-scroll"
-        >
-          <div ref={docxRef} className="office-docx-host" />
-        </div>
-      </div>
-    );
-  }
-
-  // XLSX — sheet tabs only when embedded; no filename title
-  if (kind === "xlsx") {
-    return (
-      <div
-        className={
-          "office-preview office-preview--xlsx" +
-          (embedded ? " office-preview--embedded" : "")
-        }
-      >
-        {!embedded && (
-          <div className="office-preview__bar">
-            <Tip label={name}>
-              <span className="office-preview__bar-title">
-                {name}
-              </span>
-            </Tip>
-            <div className="office-preview__bar-actions">
-              <button
-                type="button"
-                className="btn btn--ghost btn--sm"
-                onClick={() => void openExternal()}
-              >
-                {tr("office.openExternal")}
-              </button>
-            </div>
-          </div>
-        )}
-        {sheetNames.length > 1 && (
-          <div className="office-preview__sheets" role="tablist">
-            {sheetNames.map((sn, i) => (
-              <button
-                key={sn}
-                type="button"
-                role="tab"
-                className={
-                  "office-preview__sheet-tab" +
-                  (i === activeSheet ? " is-active" : "")
-                }
-                onClick={() => switchSheet(i)}
-              >
-                {sn}
-              </button>
-            ))}
-          </div>
-        )}
-        <div
-          className="office-preview__sheet-scroll"
-          dangerouslySetInnerHTML={{ __html: sheetHtml }}
+  const child = (() => {
+    if (kind === "pdf") {
+      return (
+        <OfficePdfPreview
+          buffer={load.buffer}
+          name={name}
+          locale={locale}
+          embedded={embedded}
+          onOpenExternal={() => void openExternal()}
         />
+      );
+    }
+    if (kind === "docx" || kind === "office") {
+      return (
+        <OfficeDocxPreview
+          buffer={load.buffer}
+          name={name}
+          locale={locale}
+          embedded={embedded}
+          onOpenExternal={() => void openExternal()}
+          onError={handleChildError}
+        />
+      );
+    }
+    if (kind === "xlsx") {
+      return (
+        <OfficeXlsxPreview
+          buffer={load.buffer}
+          name={name}
+          locale={locale}
+          embedded={embedded}
+          onOpenExternal={() => void openExternal()}
+          onError={handleChildError}
+        />
+      );
+    }
+    return (
+      <div className="office-preview office-preview--center">
+        <div className="office-preview__status">{tr("office.unsupported")}</div>
+        <button type="button" className="btn btn--solid" onClick={() => void openExternal()}>
+          {tr("office.openExternal")}
+        </button>
       </div>
     );
-  }
+  })();
 
   return (
-    <div className="office-preview office-preview--center">
-      <div className="office-preview__status">{tr("office.unsupported")}</div>
-      <button type="button" className="btn btn--solid" onClick={() => void openExternal()}>
-        {tr("office.openExternal")}
-      </button>
-    </div>
+    <Suspense
+      fallback={
+        <div className="office-preview office-preview--center">
+          <div className="office-preview__status">{tr("office.loading")}</div>
+          {!embedded ? (
+            <div className="office-preview__sub">{name}</div>
+          ) : null}
+        </div>
+      }
+    >
+      {child}
+    </Suspense>
   );
 }
