--- src/components/ResourceViewer.tsx
@@ -1,10 +1,0 @@
-/**
- * ResourceViewer public entry — re-exports shell + public types.
- * Implementation lives under `./resource-viewer/`.
- */
-
-export { ResourceViewer } from "./resource-viewer/ResourceViewer";
-export type {
-  ResourceOpenTarget,
-  ResourceViewerProps,
-} from "./resource-viewer/types";
