+++ src/lib/viteManualChunks.ts
@@ -9,6 +9,18 @@ export function vendorManualChunk(id: string): string | undefined {
   if (at < 0) return;
   const rest = n.slice(at + marker.length);
 
+  // React core gets its own chunk: manual chunks opt out of Rollup's auto
+  // splitting, so a feature vendor chunk that also holds react-dom becomes a
+  // static dependency of every shell that uses JSX — entry included (this is
+  // what forced the 533KB tiptap chunk into the boot preload list).
+  if (
+    rest.startsWith("react/") ||
+    rest.startsWith("react-dom/") ||
+    rest.startsWith("react-is/") ||
+    rest.startsWith("scheduler/")
+  ) {
+    return "framework";
+  }
   if (rest.startsWith("@xterm/") || rest.startsWith("xterm/")) {
     return "xterm";
   }
